Further cleanup to make a full push, file cache imported but rework of how it is managed should come around fairly soon.
Additionally, file update detection will be processed differently, and need to clean up other variables.

In other words it is very much not functioning as intended right now, but the spirit is all there.

Still very much likely room for nessisary cleanup, but cant see it until i try it!